Clean up warnings in the tests directory
Build repair.o with strict aliasing disabled because it addresses the
same data structure via two different structs at the same time.  Fix
various other, more minor warnings.

Several fixes here require more Autoconf glue to work properly, but
that's deferred to future work (such as some HAVE_* defines that we
don't probe for and the handling of non-native 64-bit integers in
some of the code).

Reviewed-on: http://gerrit.openafs.org/2060
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/tests/afscp.c
@@ -585,7 +585,7 @@ Finish:
 	double rate, size, time;
 	if (finish.tv_sec == start.tv_sec) {
 	    printf("Copied %d bytes in %d microseconds\n", filesz,
-		   finish.tv_usec - start.tv_usec);
+		   (int)(finish.tv_usec - start.tv_usec));
 	} else {
 	    printf("Copied %d bytes in %d seconds\n", filesz,
 		   (int)(finish.tv_sec - start.tv_sec));
